feat(tracing): Move `registerErrorInstrumentation` to `@sentry/core` (#7494)
